map generation algorithm done
I found I needed a lot less tracking elements than I thought.
One is a stack of tiles that might still have neighbours to explore
left. And the other is the map itself. The only way a floor tile can be
there, is because I put it there as a map creation step. So I know I
have visited that spot.

Also I'm quite happy about naming the function for drawing one specific
floor tile excavate(). It is exactly how it visually feels when we use a
timer to slow the drawing process: The tunnels are excavated into the
dungeon.

Next step would be to:
 - set start/end points for the dungeon
 - implement a level counter
 - collision with wall tiles
 - collision with map limits or resizing so there are certainly walls
   around it
 - maybe think about some kind of key-door thingy?

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>