doors and keys interactive
This was a lot more work left than I thought it would be.
Both the key and the door needed to get extracted from the normal
tilemap and put into separate scenes even.
But then playing around with the events and the separated scripts for
each interactive object type, that was quite fun.

I also like that it is almost fully instantiatable. What I mean is,
Thanks to putting doors and keys into separate scenes and using events
to orchestrate their interactions, I am almost at a point where I could
add a lot more doors and keys without additional investment into code.
The only thing missing is that the events tell each other which door and
which key instance the event is coming from.

As a next step, I really want to build a HUD, so I can show the amount
of keys found. Even if there's only one for now.

I'm thinking about another feature first: Figuring out the distance
between two end points and then putting the exit and entry points as far
away from each other as possible. That was basically the alternative to
using keys, and therefore, I now have actually already committed to the
other path. However, when debugging it is really frustrating to walk
again and again for a long time until I reach the door or key. And if I
have a sorted list of such paths, I can also choose the shortest paths
on world creation when I'm debugging, and choose the longest paths when
I'm actually playing.

All that also hints at the game loop not being fun enough. If walking is
the main thing you do, it should not be frustrating to do it.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>