pause menu working
I decided to go with the pause menu first, as it is basically a HUD
element, too. Just one that hides everything else and only appears
sometimes. Also, the tutorial for it was really good:

https://www.youtube.com/watch?v=9dlNkB-QK3c

Next task is making more HUD elements. E.g. current level and amount of
keys.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>